+++ cmd/forgeflow/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/forgeflow/forgeflow/cli"
+)
+
+func main(){if err:=cli.Execute();err!=nil{fmt.Fprintln(os.Stderr,"forgeflow:",err);os.Exit(1)}}
+